Reorganize commands and subcommands
Improve organization of subcommands and command naming. Refactor flags.

- Command-specific flag sets kept with command
- Consistent command function names
- Remove unused "register" command
- Group loadtest commands
