+++ command/flags.go
@@ -61,223 +61,6 @@ var listenP2PFlag = &cli.StringFlag{
 	Required: false,
 }
 
-var daemonFlags = []cli.Flag{
-	cacheSizeFlag,
-	listenAdminFlag,
-	listenFinderFlag,
-	listenIngestFlag,
-	listenP2PFlag,
-	&cli.BoolFlag{
-		Name:     "watch-config",
-		Usage:    "Watch for changes to config file and automatically reload",
-		EnvVars:  []string{"STORETHEINDEX_WATCH_CONFIG"},
-		Value:    true,
-		Required: false,
-	},
-}
-
-var findFlags = []cli.Flag{
-	&cli.StringSliceFlag{
-		Name:     "mh",
-		Usage:    "Specify multihash to use as indexer key, multiple OK",
-		Required: false,
-	},
-	&cli.StringSliceFlag{
-		Name:     "cid",
-		Usage:    "Specify CID to use as indexer key, multiple OK",
-		Required: false,
-	},
-	indexerHostFlag,
-	&cli.StringFlag{
-		Name:     "indexerid",
-		Usage:    "Indexer peer ID to use when protocol=libp2p",
-		Aliases:  []string{"iid"},
-		EnvVars:  []string{"INDEXER_ID"},
-		Required: false,
-	},
-	&cli.StringFlag{
-		Name:     "protocol",
-		Usage:    "Protocol to query the indexer (http, libp2p currently supported)",
-		Value:    "http",
-		Required: false,
-	},
-}
-
-var adminFreezeFlags = []cli.Flag{
-	indexerHostFlag,
-}
-
-var importFlags = []cli.Flag{
-	providerFlag,
-	&cli.StringFlag{
-		Name:     "ctxid",
-		Usage:    "Context ID of data imported",
-		Aliases:  []string{"c"},
-		Required: true,
-	},
-	&cli.StringFlag{
-		Name:     "metadata",
-		Usage:    "Bytes of opaque metadata corresponding to protocol 0",
-		Aliases:  []string{"m"},
-		Required: false,
-	},
-	fileFlag,
-	indexerHostFlag,
-}
-
-var adminPolicyFlags = []cli.Flag{
-	&cli.StringFlag{
-		Name:     "peer",
-		Usage:    "Peer ID of publisher or provider to allow or block",
-		Aliases:  []string{"p"},
-		Required: true,
-	},
-	indexerHostFlag,
-}
-
-var adminReloadConfigFlags = []cli.Flag{
-	indexerHostFlag,
-}
-
-var adminListAssignedFlags = []cli.Flag{
-	indexerHostFlag,
-}
-
-var adminListPreferredFlags = []cli.Flag{
-	indexerHostFlag,
-}
-
-var adminSyncFlags = []cli.Flag{
-	indexerHostFlag,
-	&cli.StringFlag{
-		Name:     "pubid",
-		Usage:    "Publisher peer ID",
-		Aliases:  []string{"p"},
-		Required: true,
-	},
-	&cli.StringFlag{
-		Name:  "addr",
-		Usage: "Multiaddr address of peer to sync with",
-	},
-	&cli.Int64Flag{
-		Name:  "depth",
-		Usage: "Depth limit of advertisements (distance from current) to sync. No limit if -1. Unspecified or 0 defaults to indexer config.",
-	},
-	&cli.BoolFlag{
-		Name:  "resync",
-		Usage: "Ignore the latest synced advertisement and sync advertisements as far back as the depth limit allows.",
-		Value: false,
-	},
-}
-
-var initFlags = []cli.Flag{
-	cacheSizeFlag,
-	listenAdminFlag,
-	listenFinderFlag,
-	listenIngestFlag,
-	listenP2PFlag,
-	&cli.StringFlag{
-		Name:     "store",
-		Usage:    "Type of value store (pebble, sth). Default is \"pebble\"",
-		Aliases:  []string{"s"},
-		EnvVars:  []string{"STORETHEINDEX_VALUE_STORE"},
-		Required: false,
-	},
-	&cli.BoolFlag{
-		Name:     "no-bootstrap",
-		Usage:    "Do not configure bootstrap peers",
-		EnvVars:  []string{"NO_BOOTSTRAP"},
-		Required: false,
-	},
-	&cli.StringFlag{
-		Name:     "pubsub-topic",
-		Usage:    "Subscribe to this pubsub topic to receive advertisement notification",
-		EnvVars:  []string{"STORETHEINDE_PUBSUB_TOPIC"},
-		Required: false,
-	},
-	&cli.BoolFlag{
-		Name:     "upgrade",
-		Usage:    "Upgrade the config file to the current version, saving the old config as config.prev, and ignoring other flags ",
-		Aliases:  []string{"u"},
-		Required: false,
-	},
-	&cli.BoolFlag{
-		Name:     "use-assigner",
-		Usage:    "Configure the indexer to work with an assigner service",
-		Required: false,
-	},
-	&cli.StringFlag{
-		Name:     "dhstore",
-		Usage:    "Url of DHStore for double hashed index",
-		Required: false,
-	},
-}
-
-var providersGetFlags = []cli.Flag{
-	&cli.StringFlag{
-		Name:     "provid",
-		Usage:    "Provider peer ID",
-		Aliases:  []string{"p"},
-		Required: true,
-	},
-	indexerHostFlag,
-}
-
-var providersListFlags = []cli.Flag{
-	indexerHostFlag,
-}
-
-var importProvidersFlags = []cli.Flag{
-	&cli.StringFlag{
-		Name:     "from",
-		Usage:    "Host or host:port of indexer to get providers from",
-		Aliases:  []string{"f"},
-		Required: true,
-	},
-	indexerHostFlag,
-}
-
-var registerFlags = []cli.Flag{
-	&cli.StringFlag{
-		Name:     "config",
-		Usage:    "Config file containing provider's peer ID and private key",
-		Required: true,
-	},
-	indexerHostFlag,
-	&cli.StringSliceFlag{
-		Name:     "provider-addr",
-		Usage:    "Provider address as multiaddr string, example: \"/ip4/127.0.0.1/tcp/3102\"",
-		Aliases:  []string{"pa"},
-		Required: true,
-	},
-}
-
-var statusFlags = []cli.Flag{
-	indexerHostFlag,
-}
-
-var syntheticFlags = []cli.Flag{
-	fileFlag,
-	&cli.StringFlag{
-		Name:     "type",
-		Usage:    "Type of synthetic load to generate (manifest, cidlist, car)",
-		Aliases:  []string{"t"},
-		Required: true,
-	},
-	&cli.Int64Flag{
-		Name:     "num",
-		Usage:    "Number of entries to generate",
-		Aliases:  []string{"n"},
-		Required: false,
-	},
-	&cli.Int64Flag{
-		Name:     "size",
-		Usage:    "Total size of the CIDs to generate",
-		Aliases:  []string{"s"},
-		Required: false,
-	},
-}
-
 // cliIndexer reads the indexer host from CLI flag or from config.
 func cliIndexer(cctx *cli.Context, addrType string) string {
 	idxr := cctx.String("indexer")
